requirements.txt
arrow==0.15.4
asgiref==3.4.1
astroid==2.6.6
blessed==1.16.1
certifi==2019.11.28
chardet==3.0.4
charset-normalizer==2.0.8
dj-database-url==0.5.0
Django==3.1.14
django-cors-headers==3.2.0
django-picklefield==2.0
django-q==1.3.3
django-request-logging==0.7.1
django-rest-framework==0.1.0
djangorestframework==3.11.2
enum34==1.1.10
future==0.18.2
fyle-accounting-mappings==1.7.0
fyle-integrations-platform-connector==1.17.1
fyle-rest-auth==1.1.0
gunicorn==20.0.4
idna==2.8
isort==5.10.1
lazy-object-proxy==1.6.0
mccabe==0.6.1
psycopg2-binary==2.8.4
pyjwt==2.4.0
pylint==2.7.4
python-dateutil==2.8.1
pytz==2019.3
requests==2.26.0
six==1.13.0
sqlparse==0.3.0
toml==0.10.2
Unidecode==1.1.2
urllib3==1.26.5
wcwidth==0.1.8
wrapt==1.12.1
xerosdk==0.13.0
pytest==7.1.2
pytest-cov==3.0.0
pytest-django==4.5.2
pytest-mock==3.8.2
